changed name of get_location.py to get_location_api.py for clarity of purpose and make certain some aspects of the module are still accessible. Improved validation in request_nominatim()
